add QueryModel class as ansestor of News/Job/Sale/Event Model
